+++ python/egglog/examples/bignum.py
@@ -1,0 +1,31 @@
+# mypy: disable-error-code="empty-body"
+"""
+BigNum/BigRat Example
+=====================
+"""
+
+from __future__ import annotations
+
+from egglog import *
+
+x = BigInt(-1234)
+y = BigInt.from_string("2")
+z = BigRat(x, y)
+
+assert z.numer.to_string() == "-617"
+
+
+@function
+def bignums(x: BigInt, y: BigInt) -> BigRat: ...
+
+
+egraph = EGraph()
+egraph.register(set_(bignums(x, y)).to(z))
+
+c = var("c", BigRat)
+a, b = vars_("a b", BigInt)
+egraph.check(
+    bignums(a, b) == c,
+    c.numer == a >> 1,
+    c.denom == b >> 1,
+)